requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python==0.2.0
numpy==1.24.3
pathlib==1.0.1